Made use of new generic EventPump implementation provided by Terje. Reverted original QueuingEventListener to be unchanged and added a separate QueuingEventListenerExt for custom events.
