+++ src/NUnitFramework/framework/Internal/Execution/QueuingEventListener.cs
@@ -8,7 +8,7 @@ namespace NUnit.Framework.Internal.Execution
     /// QueuingEventListener uses an EventQueue to store any
     /// events received on its EventListener interface.
     /// </summary>
-    public class QueuingEventListener : ITestListener, ITestListenerExt
+    public class QueuingEventListener : ITestListener
     {
         /// <summary>
         /// The EventQueue created and filled by this listener
@@ -40,42 +40,6 @@ namespace NUnit.Framework.Internal.Execution
         public void TestFinished(ITestResult result)
         {
             Events.Enqueue(new TestFinishedEvent(result));
-        }
-
-        /// <summary>
-        /// Called when a OneTimeSetUp has started
-        /// </summary>
-        /// <param name="test">Information about the OneTimeSetUp method that has started. Needs to be replaced with a more fitting object in the future.</param>
-        public void OneTimeSetUpStarted(ITest test)
-        {
-            Events.Enqueue(new OneTimeSetUpStartedEvent(test));
-        }
-
-        /// <summary>
-        /// Called when a OneTimeSetUp has finished
-        /// </summary>
-        /// <param name="test">Information about the OneTimeSetUp method that has finished. Needs to be replaced with a more fitting object in the future.</param>
-        public void OneTimeSetUpFinished(ITest test)
-        {
-            Events.Enqueue(new OneTimeSetUpFinishedEvent(test));
-        }
-
-        /// <summary>
-        /// Called when a OneTimeTearDown has started
-        /// </summary>
-        /// <param name="test">Information about the OneTimeTearDown method that has started. Needs to be replaced with a more fitting object in the future.</param>
-        public void OneTimeTearDownStarted(ITest test)
-        {
-            Events.Enqueue(new OneTimeTearDownStartedEvent(test));
-        }
-
-        /// <summary>
-        /// Called when a OneTimeTearDown has finished
-        /// </summary>
-        /// <param name="test">Information about the OneTimeTearDown method that has finished. Needs to be replaced with a more fitting object in the future.</param>
-        public void OneTimeTearDownFinished(ITest test)
-        {
-            Events.Enqueue(new OneTimeTearDownFinishedEvent(test));
         }
 
         /// <summary>
